move receive form in its own file
minor other fixes
